+++ examples/2services.rs
@@ -1,4 +1,4 @@
-use dtypes::DString as String;
+use dtypes::redis::DString as String;
 use std::thread;
 use std::thread::sleep;